+++ lib/document-store/canonicalize.ts
