+++ lib/document-store/current-scene.ts
@@ -19,7 +19,7 @@ function key(stageId: string): string {
 
 function resolveKv(kv?: KVStore): KVStore {
   if (kv) return kv;
-  if (typeof window === 'undefined') throw new Error('Current-scene persistence is client-only');
+  if (typeof localStorage === 'undefined') throw new Error('Current-scene persistence requires localStorage (client-only)');
   return (defaultKv ??= new BrowserKVStore());
 }
