+++ lib/document-store/validators.ts
