+++ lib/document/index.ts
@@ -49,30 +49,3 @@ export type {
   MediaTranscriptSegment,
 } from './types';
 export type { DocumentBundleResult, ParsedDocumentImage, ParsedDocumentPart } from './bundle';
-export type { AppDocument, AppDocumentOutline, AppStage } from './persistence-types';
-export {
-  canonicalizeLegacyOutline,
-  canonicalizeLegacyScene,
-  canonicalizeLegacyStage,
-} from './canonicalize';
-export { getDocumentStore, type DocumentStoreDeps } from './store';
-export {
-  accessDocument,
-  documentLockName,
-  getLegacyDocumentStore,
-  mutateDocument,
-  withDocumentLock,
-  DocumentLockUnavailableError,
-  type DocumentAccessResult,
-  type DocumentMigrationDeps,
-  type LegacyDocumentSnapshot,
-  type LegacyDocumentStore,
-} from './migration';
-export {
-  clearCurrentScene,
-  loadCurrentScene,
-  saveCurrentScene,
-  type CurrentSceneDeps,
-  type CurrentSceneValue,
-} from './current-scene';
-export { validateAppScene, validateAppStage } from './validators';
